requirement.txt
scipy
decision-tree-id3
request